Add paste statement for soils
When saving all of the SOILWAT output, the soils_v30.in file that gets
used is the one in the sw_src folder. Thus, I added code that will also
copy the soil file into that folder as well as the Stepwat_Inputs
folder, which is where the model will look for the soils_v30.in if you
are NOT saving the SOILWAT output.