removed code that deals with shifted "Year" column
There was a problem with the bmassavg.csv file produced from the C
code, such that when it was read into R the first column of data (Year)
was lost. This R code fixed that problem before the bmass output was
put into the SQLite database. That issue has been partially resolved in
the C code, so I remove the R code that fixed the issue, as it is no
longer necessary.