Few code fixes and added comments
Edited the code in the makedatabase.R file to now accommodate the
standard deviation columns for each group in the total_bmass.csv file.
Edited the Weather Assembly and Wrapper Code files to comment out code
that creates weather.in files. The default is to only use the markov
generated weather to run STEPWAT2. Added comments that explain the
StepWat.Wrapper.Code_V3.R file more clearly and cleaned up the file a
bit. Also cleaned up the main file a bit.