Correcting calculation of ppt_avg (issue #28)
In the last commit related to issue #28 I changed how PPT_avg was calculated to avoid NaN values. However, the new implementation divided by the total number of years for that day, instead of the total number of WET days for that year. This error has been corrected with this commit.