changed naming of species.type to species (issue #35)
This change has already been implemented in MAIN and needs to be updated here. species.type has been simplified to species.

In addition shat was previously called "species" in the for loop on line 18 has been changed to "sp" because species is now taken.

The indentation of this file has been corrected to meet our style standards. This will make determining where for loops end much easier.