Preventing NaN value for PPT_avg (issue #28)
A check has been added when calculating PPT_avg to avoid division by 0. Division by 0 occurs in the case where there has never been rainfall for the given DOY. This results in 0/0 which equals NaN. This check enures that the denominator is greater than 0. If it is not, PPT_avg defaults to 0.