Addressed issue #44 - added default lines for testing purposes
To expedite testing, I have added lines for climate scenarios, soils, and disturbances that can be utilized for testing purposes. These are the default settings. The full range of options currently utilized is below these default settings and can simply be uncommented out when the user wants to run simulations.