Added species column to aggregate loop, addressing issue #65
The total_sw2_daily and total_sw2_daily_slyrs Tables were still missing a species column because those data are further modified in makedatabase.R. This commit adds the species column to the aggregate functions and solves this problem.

The species column is now present in all Tables of the SQlite databases.