Overhauled structure of rgroup code (issue #87 and issue #86)
This commit changes the naming convention of rgroup files. The convention is now "rgroup.freq.x.graz.y.z.a"
where x = disturbance freqency
           y = grazing frequency
           z = grazing intensity
           a = treatment name
All file names are now stored in the 'rgroups' variable, and instead of looping through 'dist.freq', 'graz.freq', and 'graz_intensity', there is a single loop through 'rgroups'. The dst,grz,and intensity are still needed for output, so the file name string is parsed for x, y and z.

In addition, the nograz/lowgraz issue has been resolved by assigning "0" to graz_intensity if graz.freq is 0. Otherwise, graz_intensity will be the maximum value in the 'proportion_grazing' column of input.