Moved instantiation of variables out of if statements (issue #85)
The code was still failing if no subset of sites were specified in inputs. This is because treatments_vector and treatments_vector_species were both instantiated inside of if statements, but referenced later on outside of if statements. This meant that when the if statements was false, the variable was never created and when it is referenced later on the program throws a warning.
This is fixed by instantiating the variables to empty vectors outside of any conditions.
In addition, the call to change directories to STEPWAT_DIST only occured if there was a subset of sites specified. This prevented the program from finding the template files or writing to STEPWAT_DIST. This was solved by moving the command setwd("STEPWAT_DIST") outside of any condition.