Fixed space parameters not summing to one (issue #117)
After running the code line-by-line I was able to determine that the problem with space not summing to one was actually caused by not resetting the _space vectors.  These vectors store the space parameters received from the input files. The problem arises because values are added recursively to the _space vectors, but they are never removed. Therefore, on the second iteration of the for(rg in rgroups) loop they still contain values. This means that is two shrubs are declared in the first rgroup file, shrub_space = [x, y], and on the second iteration of the loop shrub_space = [x, y, z, a]. This causes that values rewritten to the file to be much smaller than they should be.
The problem is solved by moving the declarations of the _space vectors inside of the rgroups loop.