Multiple phenology changes
(issue #165) Added a rescale_phenology flag. When TRUE sxwphen.in and sxwprod_v2.in files will be generated for each climate scenario.

(issue #159) I also ensured the phenology values sum to 1 across rows.

I also rounded all of the values we write to 9 decimal places, to make sure we aren't writing scientific notation to the files (which c might or might not be able to read).