Compares phenology files
(issue #168)
I added the functionallity to calculate the difference between the generated sxwphen.in files and the original values in the spread sheet. For each sxwphen.in file a separate .csv file is generated with the results. The files are named "<generated file> -original.csv" and are located in the output directory. (the output directory is created while the script is run).