Finished comparisons for sxwprod tables (also modified .gitignore)
(issue #168)
sxwprod.in files are now compared to the master CSV files from inputs.

Three output CSV files are generated for every sxwprod.in file in the STEPWAT_DIST directory: LITTER, BIOMASS, and PCTLIVE. I decided to make three output CSVs because it was the path or least resistance and it makes each file easier to read.

I also reorganized the file struture. comparisons are now stored in directories specific to their file type, i.e. sxwprod files are stored in output/sxwprod_comparisons/ and sxwphen comparisons are stored in output/sxwphen_comparisons/.

In the .gitignore file I added STEPWAT_DIST/output. I also noticed that we forgot to change Stepwat_R_Wrapper_Parallel to R_program when we renamed the directories, so I corrected that as well.

Finally, I reogranized the comments in comparaFiles.R