Added prod graphics and some security checks
The prod graphics are generated for litter, pctlive, and biomass. For pctlive and biomass I created a graphic for each RGroup.

I also added a stopifnot statement to the top of the program to make sure we are actually in the STEPWAT_DIST directory. Otherwisethis program could be run anywhere and make a mess of the file system.

All calls to create directories are now contained in if(!file.exists()) blocks to ensure the system doesn't generate any warnings.

issue #168