Added "current" scenario to all phenology graphics
As requested, all phenology graphics now contain the the current scenario, if one exists. To determine if a current file exists I searched for the pattern "urrent", to avoid any issues with case.

Each graphic now supports 9 files plus current. If a current scenario isn't present there are only 9 files displayed. I did this to limit the amount of code and because we expect a current file to be present in almost every simulation we run.

(issue #168)