Output growing season information to STEPWAT_DIST
NOTE: Although this commit changes Main.R, it does not affect the output of the program. It only adds a side effect used for later data analysis.

This is the first half of adding growing season information to the output files generated in compareFiles.R. (issue #168)

This commit adds a boolean parameter to scale_phenology. If TRUE a list of booleans will be appended to the information returned. The booleans are TRUE if the average temperature for that month and climate scenario is above 4 degrees celsius and FALSE if it is below.

The next spet is to incorporate this information into the graphics generated in compareFiles.R