Added growing season information to phenology graphics
The growing seasons for each file are now displayed at the bottom of the graphs as lines. Where the lines are visible is the growing season. The lines are staggered to make it easy to see all of them.

The only time the lines look a little strange is when there are less than 10 files displayed, because "Current" is hard coded to it's y location.

(issue #168)