Main.R now outputs the original space values on request
(issue #168)
In order to compare the generated rgroup.in files we must know the original space values from inputs. Since we input multiple space values that apply to a subset of all sites, we need to output the space values prescaling to compare them.

The alternative is to parse the input CSV a second time to derive the space values. This seems like a waste of time, so I chose to create an output file with exactly what we need.