Removed growing season code from compareFiles.R and fixed bug.
issue #184.
Since we no longer have growing seasons as an input to the program there is no need for the code in compareFiles.R to add it to the graphics.

I also found a bug in commit 5acfc38. I removed the lines that set phenology, pctlive, littler, and biomass to the newly scaled values. I meant to only remove the transpose function, so I added the lines back.

Lastly I added site_latitude back to the documentation for scale_phenology. We still need site_latitude to generate the mean monthly temperature.